Files/LB3-Code/requirements.txt
Flask
Flask_login
Flask-Cors
mysql-connector-python
requests
uuid
datetime  
camunda-external-task-client-python
pycamunda